requirements.txt
absl-py==2.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2024.6.1
grpcio==1.74.0
humanfriendly==10.0
imageio==2.37.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
onnx==1.18.0
onnxruntime==1.22.1
onnxruntime-directml==1.22.0
packaging==25.0
pandas==2.3.2
pillow==11.0.0
protobuf==6.32.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
